ai-fastapi-rag/auth/auth.py
from fastapi import APIRouter, HTTPException, Depends, Form
from pydantic import BaseModel
from passlib.context import CryptContext
from jose import jwt, JWTError
from dotenv import load_dotenv
import os


load_dotenv()
SECRET_KEY = os.getenv("SECRET_KEY",)