ai-fastapi-rag/requirement.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
PyMuPDF
langchain
chromadb